requirements2.txt
pip >= 21.0
tensorflow
keras==2.7.0
Pillow